fix bug: identity解析时类型错误

+++ src/main/java/com/beecloud/service/MqttServiceImpl.java
@@ -120,7 +120,7 @@ public class MqttServiceImpl implements MqttService{
         Gson gson = new Gson();
         AuthObject authObject = gson.fromJson(identity,AuthObject.class);
         long identityCode = authObject.getIdentityCode();
-        int tobeReplace = JsonPath.parse(message).read("$.identity.identityCode");
+        long tobeReplace = JsonPath.parse(message).read("$.identity.identityCode");
         message = message.replace(String.valueOf(tobeReplace),String.valueOf(identityCode));
         SendMessageObject sendMessageObject = Util.transJsonToAbstractMessage(message);
         Util.sendMessageByVin(authObject.getVin().trim(),thread_Group_function,sendMessageObject);
